use double precision for airmass calculation

+++ src/core/RefractionExtinction.hpp
@@ -99,7 +99,7 @@ private:
 	//! A problem ist that refraction depends on air pressure and temperature, but Young's formula assumes T=15C, p=1013.25mbar.
 	//! So, it seems better to compute refraction first, and then use the Rozenberg formula here.
 	//! Rozenberg is infinite at Z=92.17 deg, Young at Z=93.6 deg, so this function RETURNS SUBHORIZONTAL_AIRMASS BELOW -2 DEGREES!
-	float airmass(float cosZ, const bool apparent_z=true) const;
+	double airmass(float cosZ, const bool apparent_z=true) const;
 
 	//! k, magnitudes/airmass, in [0.00, ... 1.00], (default 0.20).
 	float ext_coeff;
